feat: add attic-hs and vira-types packages; refactor CI environment and pipeline types
