+++ cabal.project
@@ -1,5 +1,7 @@
 packages:
   packages/vira
+  packages/vira-types
+  packages/attic-hs
   packages/tail
   packages/warp-tls-simple
   packages/htmx-lucid-contrib
