+++ packages/vira-types/vira-types.cabal
@@ -1,0 +1,82 @@
+cabal-version: 2.2
+
+-- This file has been generated from package.yaml by hpack version 0.37.0.
+--
+-- see: https://github.com/sol/hpack
+
+name:           vira-types
+version:        0.1.0.0
+synopsis:       Vira shared types
+description:    Core types and data structures for vira
+category:       Web
+homepage:       https://github.com/juspay/vira
+author:         Sridhar Ratnakumar
+maintainer:     srid@srid.ca
+copyright:      2024 Juspay
+license:        AGPL-3.0-or-later
+build-type:     Simple
+
+library
+  exposed-modules:
+      Vira.CI.Environment.Type
+      Vira.CI.Pipeline.Type
+      Vira.State.Type
+  other-modules:
+      Paths_vira_types
+  autogen-modules:
+      Paths_vira_types
+  hs-source-dirs:
+      src
+  default-extensions:
+      DataKinds
+      DeriveDataTypeable
+      DeriveGeneric
+      DeriveTraversable
+      DerivingStrategies
+      DerivingVia
+      ExplicitForAll
+      FlexibleContexts
+      FlexibleInstances
+      GeneralizedNewtypeDeriving
+      ImportQualifiedPost
+      LambdaCase
+      MultiParamTypeClasses
+      MultiWayIf
+      NamedFieldPuns
+      NoStarIsType
+      NumericUnderscores
+      OverloadedStrings
+      ScopedTypeVariables
+      StrictData
+      TypeApplications
+      TypeFamilies
+      TypeOperators
+      TypeSynonymInstances
+      ViewPatterns
+  ghc-options: -Wall -optP-Wno-nonportable-include-path -Wincomplete-record-updates -Wincomplete-uni-patterns -Wmissing-deriving-strategies -Wunused-foralls -fprint-explicit-foralls -fprint-explicit-kinds
+  build-depends:
+      aeson
+    , attic-hs
+    , base ==4.*
+    , base16-bytestring
+    , bytestring
+    , containers
+    , data-default
+    , deriving-aeson
+    , effectful
+    , effectful-core
+    , git-effectful
+    , http-api-data
+    , ixset-typed
+    , optics-core
+    , optics-th
+    , relude >=1.0
+    , safecopy
+    , servant
+    , text
+    , time
+  mixins:
+      base hiding (Prelude)
+    , relude (Relude as Prelude, Relude.Container.One)
+    , relude 
+  default-language: GHC2021
