+++ packages/vira/vira.cabal
@@ -42,7 +42,6 @@ library
       Vira.App.Stack
       Vira.CI.Environment
       Vira.CI.Pipeline
-      Vira.Lib.Attic
       Vira.Lib.Cachix
       Vira.Lib.Logging
       Vira.Lib.Omnix
@@ -60,7 +59,6 @@ library
       Vira.State.Acid
       Vira.State.Core
       Vira.State.JSON
-      Vira.State.Type
       Vira.Stream.Log
       Vira.Stream.Refresh
       Vira.Supervisor.Core
@@ -112,6 +110,7 @@ library
     , acid-state
     , aeson
     , async
+    , attic-hs
     , base ==4.*
     , base16-bytestring
     , binary
@@ -165,6 +164,7 @@ library
     , tail
     , text
     , time
+    , vira-types
     , wai
     , wai-extra
     , wai-middleware-static
@@ -218,6 +218,7 @@ executable vira
     , acid-state
     , aeson
     , async
+    , attic-hs
     , base ==4.*
     , base16-bytestring
     , binary
@@ -272,6 +273,7 @@ executable vira
     , text
     , time
     , vira
+    , vira-types
     , wai
     , wai-extra
     , wai-middleware-static
@@ -326,6 +328,7 @@ test-suite vira-tests
     , acid-state
     , aeson
     , async
+    , attic-hs
     , base ==4.*
     , base16-bytestring
     , binary
@@ -380,6 +383,7 @@ test-suite vira-tests
     , text
     , time
     , vira
+    , vira-types
     , wai
     , wai-extra
     , wai-middleware-static
